+++ RELEASE_NOTES.md
@@ -2,9 +2,10 @@
 - [KOGITO-4245](https://issues.redhat.com/browse/KOGITO-4245) - Use Infinispan CR spec data instead of status to fetch credentials Secret 
 - [KOGITO-2317](https://issues.redhat.com/browse/KOGITO-2317) - Unify Kogito Quarkus and Spring Images
 - [KOGITO-4059](https://issues.redhat.com/browse/KOGITO-4059) Refactor Kogito operator into internal and core component
+- [KOGITO-4173](https://issues.redhat.com/browse/KOGITO-4173) - Adjust default probe URL for Quarkus
 
 ## Bug Fixes
 - [KOGITO-4332](https://issues.redhat.com/browse/KOGITO-4332) - Operator does not have rights on mongodb crds
+- [KOGITO-4210](https://issues.redhat.com/browse/KOGITO-4210) - Fix probe config for Quarkus deployments
 
 ## Known Issues
-